requirements.txt
numpy
parameterized
pillow